requirements.txt
# TeacherManager requirements.txt
# Synced with pyproject.toml via: poetry export --without-hashes -f requirements.txt
# Generated for Python 3.12

# Core FastAPI & Web
fastapi
uvicorn[standard]
starlette
pydantic
pydantic-settings
pydantic-core
python-multipart

# Database & Migrations
sqlalchemy
asyncpg
alembic
greenlet

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
cryptography
bcrypt

# PDF Generation
weasyprint
tinycss2
fonttools

# Caching & Sessions
redis

# APM & Monitoring
sentry-sdk

# Async & File Handling
aiofiles
httpx

# Environment
python-dotenv

# Type Hints (optional, for IDE/mypy)
typing-extensions
